requirements.txt
numpy
pandas
jupyter
torch
Pillow
torchvision
click
timm
scipy
opencv-python
scikit-learn
scikit-image
tensorboard
matplotlib
wandb
python-dotenv
tqdm